Recommended changes (post review)
